Move the METS mdSec from the package METS into the API
The meta-data can be stored as part of METS in different METS mdSec
areas. However, an API class should not reference a class from a low
package. Therefore, the enumeration of the meta-data sections in METS is
moved to the API class.
