+++ Kitodo-DataFormat/src/main/java/org/kitodo/dataformat/access/MetadataEntry.java
@@ -11,7 +11,7 @@
 
 package org.kitodo.dataformat.access;
 
-import org.kitodo.api.dataformat.mets.MdSec;
+import org.kitodo.api.MdSec;
 import org.kitodo.api.dataformat.mets.MetadataXmlElementAccessInterface;
 import org.kitodo.dataformat.metskitodo.MetadataType;
